+++ generators/app/templates/src/styles/landing.js
@@ -1,5 +1,0 @@
-import { css } from 'react-emotion';
-
-export default css`
-  text-align: center;
-`;